# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: ReqFlow Procurement Workflow API
servers:
  - url: /api
...
paths:
...
  /requisitions/{id}/submit:
    post:
      operationId: submitRequisition
      tags: [workflow]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
  /requisitions/{id}/checker1-review:
    post:
      operationId: checker1Review
      tags: [workflow]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckerReviewInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
  /requisitions/{id}/checker2-review:
    post:
      operationId: checker2Review
      tags: [workflow]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckerReviewInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
  /requisitions/{id}/approve:
    post:
      operationId: approveRequisition
      tags: [workflow]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproverActionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
  /requisitions/{id}/reject:
    post:
      operationId: rejectRequisition
      tags: [workflow]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproverActionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
  /requisitions/{id}/assign:
    post:
      operationId: assignRequisition
      tags: [workflow]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
  /requisitions/{id}/hold:
    post:
      operationId: holdRequisition
      tags: [workflow]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HoldInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
  /requisitions/{id}/resume:
    post:
      operationId: resumeRequisition
      tags: [workflow]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
  /requisitions/{id}/complete:
    post:
      operationId: completeRequisition
      tags: [workflow]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
        "400":
          description: No approval note linked yet
  /requisitions/{id}/approval-notes:
    get:
      operationId: listApprovalNotes
      tags: [workflow]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalNote"
    post:
      operationId: addApprovalNote
      tags: [workflow]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalNote"
  /requisitions/{id}/approval-notes/{noteId}:
    delete:
      operationId: deleteApprovalNote
      tags: [workflow]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Requisition:
      type: object
      required: [id, ref_number, raised_by_name, requisition_date, status, priority, created_at, updated_at]
      properties:
        id:
          type: integer
        ref_number:
          type: string
        project_id:
          type: ["integer", "null"]
        project_name:
          type: ["string", "null"]
        site_id:
          type: ["integer", "null"]
        site_name:
          type: ["string", "null"]
        raised_by_id:
          type: ["integer", "null"]
        raised_by_name:
          type: string
        requisition_date:
          type: string
        priority:
          type: string
        purpose:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
        checker1_id:
          type: ["integer", "null"]
        checker1_name:
          type: ["string", "null"]
        checker1_suggestion:
          type: ["string", "null"]
        checker1_reviewed_at:
          type: ["string", "null"]
        checker1_status:
          type: ["string", "null"]
        checker2_id:
          type: ["integer", "null"]
        checker2_name:
          type: ["string", "null"]
        checker2_suggestion:
          type: ["string", "null"]
        checker2_reviewed_at:
          type: ["string", "null"]
        checker2_status:
          type: ["string", "null"]
        approver_id:
          type: ["integer", "null"]
        approver_name:
          type: ["string", "null"]
        approver_suggestion:
          type: ["string", "null"]
        approver_reviewed_at:
          type: ["string", "null"]
        approver_status:
          type: ["string", "null"]
        hold_reason:
          type: ["string", "null"]
        held_at:
          type: ["string", "null"]
        pre_hold_status:
          type: ["string", "null"]
        purchase_head_id:
          type: ["integer", "null"]
        purchase_head_name:
          type: ["string", "null"]
        assigned_to_id:
          type: ["integer", "null"]
        assigned_to_name:
          type: ["string", "null"]
        assigned_at:
          type: ["string", "null"]
        submitted_at:
          type: ["string", "null"]
        approved_at:
          type: ["string", "null"]
        completed_at:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
        item_count:
          type: ["integer", "null"]
        total_expected_cost:
          type: ["number", "null"]
...
    SubmitInput:
      type: object
      required: [checker1_id, checker1_name, checker2_id, checker2_name, approver_id, approver_name]
      properties:
        checker1_id:
          type: integer
        checker1_name:
          type: string
        checker2_id:
          type: integer
        checker2_name:
          type: string
        approver_id:
          type: integer
        approver_name:
          type: string
    CheckerReviewInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
        suggestion:
          type: string
        checker_name:
          type: string
    ApproverActionInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
        suggestion:
          type: string
        approver_name:
          type: string
    AssignInput:
      type: object
      required: [assigned_to_id, assigned_to_name]
      properties:
        assigned_to_id:
          type: integer
        assigned_to_name:
          type: string
...
    ApprovalNote:
      type: object
      required: [id, requisition_id, note_number, created_at]
      properties:
        id:
          type: integer
        requisition_id:
          type: integer
        note_number:
          type: string
        created_by_name:
          type: ["string", "null"]
        created_at:
          type: string
    ApprovalNoteInput:
      type: object
      required: [note_number]
      properties:
        note_number:
          type: string
    HoldInput:
      type: object
      properties:
        reason:
          type: string
...